v1

+++ samples-recipes/as2.sh
@@ -8,6 +8,7 @@ aws_access_key_id = ${SITE_KEY}
 aws_secret_access_key = ${SITE_KEY_SECRET}
 EOL
 
-aws s3 rm s3://test-bucket4569/master-builds/latest --recursive
-
+    if [ "$TRAVIS_PULL_REQUEST" = "false" ] ; then    
+        aws s3 rm s3://test-bucket4569/master-builds/latest --recursive
+    fi
